+++ lib/middlewares/static.ts
@@ -11,17 +11,19 @@ export const staticFolder = (folderPath: string = 'public') => {
       const root = path.normalize(path.resolve(path.join(process.cwd(), folderPath)));
       const extension = path.extname(req.url).slice(1);
 
-      if (req.url === '/') req.url = '/index.html';
+      let fileName = req.url
+
+      if (fileName === '/') fileName = '/index.html';
       else if (!extension) {
         try {
-          fs.accessSync(path.join(root, req.url + '.html'), fs.constants.F_OK);
-          req.url = req.url + '.html';
+          fs.accessSync(path.join(root, fileName + '.html'), fs.constants.F_OK);
+          fileName = fileName + '.html';
         } catch (e) {
-          req.url = path.join(req.url, 'index.html');
+          fileName = path.join(fileName, 'index.html');
         }
       }
 
-      const filePath = path.normalize(path.join(process.cwd(), folderPath, req.url));
+      const filePath = path.normalize(path.join(process.cwd(), folderPath, fileName));
 
       if (!path.normalize(path.resolve(filePath)).startsWith(root)) return resolve();